+++ atlasdb-config/src/main/java/com/palantir/atlasdb/config/AtlasDbRuntimeConfig.java
@@ -23,7 +23,6 @@ import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import com.palantir.atlasdb.AtlasDbConstants;
 import com.palantir.atlasdb.compact.CompactorConfig;
-import com.palantir.atlasdb.qos.config.QosClientConfig;
 import com.palantir.atlasdb.spi.KeyValueServiceRuntimeConfig;
 import com.palantir.atlasdb.stream.StreamStorePersistenceConfiguration;
 import com.palantir.atlasdb.transaction.ImmutableTransactionConfig;
@@ -77,11 +76,6 @@ public abstract class AtlasDbRuntimeConfig {
     @Value.Default
     public long getTimestampCacheSize() {
         return AtlasDbConstants.DEFAULT_TIMESTAMP_CACHE_SIZE;
-    }
-
-    @Value.Default
-    public QosClientConfig qos() {
-        return QosClientConfig.DEFAULT;
     }
 
     public abstract Optional<KeyValueServiceRuntimeConfig> keyValueService();