+++ atlasdb-container-test-utils/src/main/java/com/palantir/atlasdb/containers/UninitializedCassandraResource.java
@@ -26,7 +26,6 @@ import com.palantir.atlasdb.cassandra.CassandraKeyValueServiceRuntimeConfig;
 import com.palantir.atlasdb.cassandra.CassandraMutationTimestampProviders;
 import com.palantir.atlasdb.keyvalue.api.KeyValueService;
 import com.palantir.atlasdb.keyvalue.cassandra.CassandraKeyValueServiceImpl;
-import com.palantir.atlasdb.qos.FakeQosClient;
 import com.palantir.atlasdb.util.MetricsManagers;
 import com.palantir.common.base.Throwables;
 
@@ -39,8 +38,7 @@ public class UninitializedCassandraResource extends ExternalResource {
             CassandraKeyValueServiceRuntimeConfig::getDefault,
             CassandraContainer.LEADER_CONFIG,
             CassandraMutationTimestampProviders.legacyModeForTestsOnly(),
-            true,
-            FakeQosClient.INSTANCE);
+            true);
     private AtomicBoolean initialized = new AtomicBoolean(false);
 
     public UninitializedCassandraResource(Class<?> classToSaveLogsFor) {